+++ lib/src/authsignal_qr.dart
@@ -3,10 +3,10 @@ import 'package:flutter/services.dart';
 
 import 'types.dart';
 
-class AuthsignalPush {
+class AuthsignalQr {
   final AsyncCallback initCheck;
 
-  AuthsignalPush({required this.initCheck});
+  AuthsignalQr({required this.initCheck});
 
   @visibleForTesting
   final methodChannel = const MethodChannel('authsignal');
@@ -16,7 +16,7 @@ class AuthsignalPush {
 
     try {
       final data = await methodChannel
-          .invokeMapMethod<String, dynamic>('push.getCredential');
+          .invokeMapMethod<String, dynamic>('qr.getCredential');
 
       if (data != null) {
         return AuthsignalResponse(data: AppCredential.fromMap(data));
@@ -35,7 +35,7 @@ class AuthsignalPush {
 
     try {
       final data = await methodChannel
-          .invokeMapMethod<String, dynamic>('push.addCredential', arguments);
+          .invokeMapMethod<String, dynamic>('qr.addCredential', arguments);
 
       if (data != null) {
         return AuthsignalResponse(data: AppCredential.fromMap(data));
@@ -52,7 +52,7 @@ class AuthsignalPush {
 
     try {
       final data =
-          await methodChannel.invokeMethod<bool>('push.removeCredential');
+          await methodChannel.invokeMethod<bool>('qr.removeCredential');
 
       if (data != null) {
         return AuthsignalResponse(data: data);
@@ -64,15 +64,18 @@ class AuthsignalPush {
     }
   }
 
-  Future<AuthsignalResponse<AppChallenge?>> getChallenge() async {
+  Future<AuthsignalResponse<ClaimChallengeResponse>> claimChallenge(
+      String challengeId) async {
     await initCheck();
+
+    var arguments = <String, dynamic>{'challengeId': challengeId};
 
     try {
       final data = await methodChannel
-          .invokeMapMethod<String, dynamic>('push.getChallenge');
+          .invokeMapMethod<String, dynamic>('qr.claimChallenge', arguments);
 
       if (data != null) {
-        return AuthsignalResponse(data: AppChallenge.fromMap(data));
+        return AuthsignalResponse(data: ClaimChallengeResponse.fromMap(data));
       } else {
         return AuthsignalResponse(data: null);
       }
@@ -81,10 +84,11 @@ class AuthsignalPush {
     }
   }
 
-  Future<AuthsignalResponse<bool>> updateChallenge(
-      {required String challengeId,
-      required bool approved,
-      String? verificationCode}) async {
+  Future<AuthsignalResponse<bool>> updateChallenge({
+    required String challengeId,
+    required bool approved,
+    String? verificationCode,
+  }) async {
     await initCheck();
 
     var arguments = <String, dynamic>{
@@ -96,7 +100,7 @@ class AuthsignalPush {
 
     try {
       final data = await methodChannel.invokeMethod<bool>(
-          'push.updateChallenge', arguments);
+          'qr.updateChallenge', arguments);
 
       if (data != null) {
         return AuthsignalResponse(data: data);
@@ -108,3 +112,4 @@ class AuthsignalPush {
     }
   }
 }
+